refactor(chat): extract native user position policy
